// ui/src/api/generated/model/checklistGroupBlock.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */
import type { ChecklistItemResponse } from './checklistItemResponse';

export interface ChecklistGroupBlock {
  group_name: string;
  items: ChecklistItemResponse[];
}